requirements.txt
pandas==2.2.1
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.1.post1
requests==2.31.0
python-dotenv==1.0.1
pyTelegramBotAPI==4.15.4
